requirements.txt
openai==1.41.1
numpy==2.1.0
pymongo==4.8.0
fastapi[standard]==0.112.1
dnspython==2.6.1
gradio==4.41.0
python-dotenv